feat: fix Project attribute validation
- added interceptor to handle the conversion of camelCase to snake_case keys
- added case converter utils
- fixed speed dial position on projects page
- added auth interceptor that adds bearer token to all requests if it is in localstorage.

+++ backend/console/migrations/m260125_165015_create_project_table.php
@@ -25,7 +25,7 @@ class m260125_165015_create_project_table extends Migration
             'visibility' => $this->string(20)->notNull()->defaultValue(Project::VISIBILITY_PUBLIC),
             'priority' => $this->integer()->notNull()->defaultValue(Project::PRIORITY_MEDIUM),
             'color' => $this->string(7)->null(),
-            'progress' => $this->integer()->notNull()->defaultValue(0),
+            'progress' => $this->float(2)->notNull()->defaultValue(0),
             'budget' => $this->decimal(10, 2)->null(),
             'created_at' => $this->integer()->notNull(),
             'updated_at' => $this->integer()->notNull(),
